+++ docs/intro.md
@@ -31,7 +31,7 @@ The react app was generated by create-react-app, several years ago. We're in the
 ### Things to know
 
 * There are Nav components and Map components, they communicate using redux.
-* We get tilesets from Mapbox Studio that we reference in components/map/MapLayers
+* We get tilesets from Mapbox Studio that we reference in `src/components/map/MapLandDataLayers`
 * Our base tileset uses ordnance survey data
 
 ### Editing the Styles
