+++ src/assets/styles/_tooltips.scss
@@ -115,7 +115,7 @@
   }
 }
 
-.tooltip-menu-main {
+.tooltip-static-site-menu {
   z-index: 100001 !important;
 }
