+++ src/components/left-pane/LandDataLayerToggle.js
@@ -5,7 +5,9 @@ import LeftPaneToggle from './LeftPaneToggle';
 
 const LandDataLayerToggle = ({ title, layerId, draggable = false }) => {
     const dispatch = useDispatch();
-    const activeLayers = useSelector((state) => state.mapLayers.landDataLayers);
+    const activeLayers = useSelector(
+      (state) => state.landDataLayers.landDataLayers
+    );
 
     const onToggle = () => {
         dispatch({ type: "TOGGLE_LAND_DATA_LAYER", payload: layerId });
